+++ src/main/java/ch/agridata/agreement/dto/DataRequestUpdateDto.java
@@ -1,6 +1,7 @@
 package ch.agridata.agreement.dto;
 
 import ch.agridata.common.utils.ValidationSchemaGenerator;
+import ch.agridata.common.validation.IsoCountryCode;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
@@ -44,13 +45,11 @@ public record DataRequestUpdateDto(
     DataRequestPurposeDto purpose,
 
     @Schema(
-        description = "Shorter name of the data consumer defined by the data consumer used when displaying the request to the producer",
-        examples = {"Bio Suisse"}
+        description = "Shorter name of the data consumer defined by the data consumer used when displaying the request to the producer"
     )
-    @Size(max = 255)
-    @Size(min = 3, max = 255, groups = ValidationSchemaGenerator.Submit.class)
     @NotNull(groups = ValidationSchemaGenerator.Submit.class)
-    String dataConsumerDisplayName,
+    @Valid
+    DataRequestConsumerDisplayNameDto dataConsumerDisplayName,
 
     @Schema(
         description = "City of the data consumer",
@@ -80,12 +79,13 @@ public record DataRequestUpdateDto(
     String dataConsumerStreet,
 
     @Schema(
-        description = "2 letter country code of the data consumer",
+        description = "2 letter valid ISO 3166-1 alpha-2 country code of the data consumer",
         examples = {"CH"}
     )
     @Size(max = 2)
     @Size(min = 2, max = 2, groups = ValidationSchemaGenerator.Submit.class)
     @NotNull(groups = ValidationSchemaGenerator.Submit.class)
+    @IsoCountryCode
     String dataConsumerCountry,
 
     @Schema(