+++ src/test/java/integration/agreement/ConsentRequestAggregationTest.java
@@ -184,7 +184,8 @@ class ConsentRequestAggregationTest {
                                     'fr', 'Surveillance légalement prescrite de la santé animale et prévention des épizooties.',
                                     'it', 'Sorveglianza legalmente prescritta della salute degli animali e prevenzione delle epizoozie.'),
                  'ACTIVE', '2025-03-11 09:27:55'::timestamp, 'Bern', 'CH',
-                 'Bundesamt für Lebensmittelsicherheit und Veterinärwesen BLV', 'BLV', 'Schwarzenburgstrasse 155', '3003',
+                 'Bundesamt für Lebensmittelsicherheit und Veterinärwesen BLV',
+                 jsonb_build_object('de', 'BLV', 'fr', 'OSAV', 'it', 'USAV'), 'Schwarzenburgstrasse 155', '3003',
                  '+41 58 463 30 33', 'info@blv.admin.ch', 'Alle Tierhalter',
                  'COLLECTIVE_SIGNATURE', 'COLLECTIVE_SIGNATURE', '[]'::jsonb)
               """)
@@ -196,7 +197,7 @@ class ConsentRequestAggregationTest {
                 (id, archived, created_at, modified_at, data_producer_uid, data_producer_bur, state_code, data_request_id)
               VALUES
                 (:uidBasedId, false, NOW(), NOW(), :uid, NULL, 'LEGALLY_PERMITTED', :dataRequestId),
-                (:burBasedId, false, NOW(), NOW(), :uid, '99910099', 'LEGALLY_PERMITTED', :dataRequestId)
+                (:burBasedId, false, NOW(), NOW(), :uid, 'A99910099', 'LEGALLY_PERMITTED', :dataRequestId)
               """)
           .setParameter("uidBasedId", UUID.randomUUID())
           .setParameter("burBasedId", UUID.randomUUID())